两个方向的退网逻辑整理完成，编译通过，未测试

+++ TinyNet/TinyServer.h
@@ -60,7 +60,7 @@ public:
 	// 组网
 	bool OnJoin(const TinyMessage& msg);
 	
-    bool Disjoin(TinyMessage& msg,uint crc) const;
+    bool Disjoin(TinyMessage& msg, ushort crc) const;
 	bool OnDisjoin(const TinyMessage& msg);
 
 	// 心跳
